[add] basic styling for login/signup

+++ client/app/components/authentication/authController.js
@@ -17,14 +17,19 @@ angular.module('app.auth', [])
 		};
 
 		$scope.signup = function(){
-			Auth.signup($scope.user)
-			.then(function (token){
-				$window.localStorage.setItem('com.bomb', token);
-				$location.path('/home');
-			})
-			.catch(function (err){
-				console.log(err);
-				$scope.user.err = true;
-			});
-		};
-	})
+			if($scope.user.password === $scope.user.confirm){
+				$scope.check = false;
+				Auth.signup($scope.user)
+				.then(function (token){
+						$window.localStorage.setItem('com.bomb', token);
+						$location.path('/home');
+					})
+				.catch(function (err){
+						console.log(err);
+						$scope.user.err = true;
+					});
+				} else {
+					$scope.check = true;
+				}
+		}
+	});
